backend/requirements.txt
Django==6.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
opencv-python==4.8.1.78
torch==2.1.0
torchvision==0.16.0
ultralytics==8.0.230
Pillow==10.1.0
numpy==1.24.3
